Fixes #2664: The apoc.bolt.* procedures are not recognized
