+++ docs/asciidoc/modules/ROOT/pages/database-integration/bolt-neo4j.adoc
@@ -74,6 +74,16 @@ In addition, the `apoc.bolt.load.fromLocal` can have:
 * `readOnly`: default false. To execute or not, read-only statements.  
 * `localParams`: to put optional parameters to local cypher statement
 
+== Install Dependencies
+
+The Bolt procedures have dependencies on a client library that is not included in the APOC Library.
+
+You can download it from https://repo1.maven.org/maven2/org/neo4j/driver/neo4j-java-driver/4.0.0/neo4j-java-driver-4.0.0.jar[mvnrepository]
+or https://github.com/neo4j-contrib/neo4j-apoc-procedures/releases/download/{apoc-release}/apoc-bolt-dependencies-{apoc-release}.jar[apoc repository].
+Once that file is downloaded, it should be placed in the `plugins` directory and the Neo4j Server restarted.
+
+
+
 == Driver configuration
 
 To set the configuration of the Driver, you can add the parameter `driverConfig` in the config.
