+++ extra-dependencies/bolt/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Tue Feb 06 14:27:44 CET 2018
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-6.1.1-bin.zip
