+++ extra-dependencies/settings.gradle
@@ -6,3 +6,4 @@ include('couchbase')
 include('xls')
 include('redis')
 include('hadoop')
+include('bolt')
